+++ src/Entities/ChatPhoto.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * This file is part of the TelegramBot package.
+ *
+ * (c) Avtandil Kikabidze aka LONGMAN <akalongman@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Longman\TelegramBot\Entities;
+
+/**
+ * Class ChatPhoto
+ *
+ * @link https://core.telegram.org/bots/api#chatphoto
+ *
+ * @method string getSmallFileId() Unique file identifier of small(160x160) chat photo. This file_id can be used only for photo download.
+ * @method string getBigFileId()   Unique file identifier of big(640x640) chat photo. This file_id can be used only for photo download.
+ */
+class ChatPhoto extends Entity
+{
+
+}